requirements.txt
Django==1.8.4
Markdown==2.6.2
MySQL-python==1.2.5
PyYAML==3.11
cached-property==1.2.0
dj-database-url==0.3.0
dj-static==0.0.6
django-admin-bootstrapped==2.5.5
django-auto-rest-project==0.1.1
django-braces==1.8.1
django-cors-headers==1.1.0
django-drf-file-generator==0.1.0
django-filter==0.11.0
django-jet==0.0.9
django-oauth-toolkit==0.9.0
django-rest-auth==0.5.0
django-rest-swagger==0.3.4
django-toolbelt==0.0.1
django-url-filter==0.2.0
djangorestframework==3.2.3
enum34==1.0.4
funcsigs==0.4
gunicorn==19.3.0
mock==1.3.0
oauthlib==1.0.1
pbr==1.8.0
psycopg2==2.6.1
python-memcached==1.57
simplejson==3.8.0
six==1.9.0
static3==0.6.1
wsgiref==0.1.2
yet-another-django-profiler==1.0.0
